ecommerce/requirements.txt
django==5.1.3      
djangorestframework==3.15.2
pyyaml==6.0.2
requests==2.32.3
django-cors-headers==4.4.0
